py/apps/app-template/pyproject.toml
[project]
name = "app-template"
version = "0.1.0"
description = "Template FastAPI backend for agentic apps"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.28.0",
  "structlog>=24.4.0",
  "opentelemetry-api>=1.29.0",
  "opentelemetry-sdk>=1.29.0",
  "opentelemetry-instrumentation-fastapi>=0.50b0",
  "opentelemetry-exporter-otlp-proto-http>=1.29.0",
  "azure-monitor-opentelemetry-exporter>=1.0.0b34",
  "azure-cosmos>=4.9.0",
  "foundrykit",
  "agentkit",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
foundrykit = { workspace = true }
agentkit = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "models", "services", "tools"]
